Support complex object keys and array items
Allow multiple classes to be specified in object keys or array items
and ensure that they are properly bemified. Also make sure that we
still throw if we receive multiple elements in a complex key/item.

One particular use case for this might be `className` in React. You
could pass a multi-class `className` down through some components and
then call `bemify` with `className` as an item in the array or a key
within an object, and if any elements or modifiers were specified in
`className`, they should be applied correctly.